[6542] Extract the draggable feature from the palette
Bug: https://github.com/eclipse-sirius/sirius-web/issues/6542
