+++ packages/diagrams/frontend/sirius-components-diagrams/src/renderer/palette/Palette.tsx
@@ -26,12 +26,8 @@ import IconButton from '@mui/material/IconButton';
 import Paper from '@mui/material/Paper';
 import { Theme, useTheme } from '@mui/material/styles';
 import Tooltip from '@mui/material/Tooltip';
-import { Edge, Node, useStoreApi } from '@xyflow/react';
-import React, { useEffect, useState } from 'react';
-import Draggable, { DraggableData } from 'react-draggable';
+import React, { useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { EdgeData, NodeData } from '../DiagramRenderer.types';
-import { useGetUpdatedModalPosition } from '../hooks/useGetUpdatedModalPosition';
 import {
   GQLPaletteDivider,
   GQLPaletteEntry,
@@ -55,96 +51,55 @@ export const isTool = (entry: GQLPaletteEntry): entry is GQLTool => !isPaletteDi
 
 const paletteWidth = 200;
 
-export const Palette = ({
-  x: paletteX,
-  y: paletteY,
-  diagramElementIds,
-  palette,
-  onToolClick,
-  onClose,
-  paletteToolListExtensions,
-}: PaletteProps) => {
-  const { domNode } = useStoreApi<Node<NodeData>, Edge<EdgeData>>().getState();
-  const { getUpdatedModalPosition, getUpdatedBounds } = useGetUpdatedModalPosition();
-  const nodeRef = React.useRef<HTMLDivElement>(null);
-  const theme: Theme = useTheme();
-  const { t } = useTranslation('sirius-components-diagrams', { keyPrefix: 'palette' });
-  const [state, setState] = useState<PaletteState>({
-    searchToolValue: '',
-    controlledPosition: { x: paletteX, y: paletteY },
-  });
+export const Palette = React.forwardRef<HTMLDivElement, PaletteProps & React.HTMLAttributes<HTMLDivElement>>(
+  (
+    {
+      diagramElementIds,
+      onClose,
+      onToolClick,
+      palette,
+      paletteToolListExtensions,
+      x: paletteX,
+      y: paletteY,
+      ...remainingProps
+    },
+    ref
+  ) => {
+    const nodeRef = React.useRef<HTMLDivElement>(null);
+    const theme: Theme = useTheme();
 
-  useEffect(() => {
-    setState((prevState) => {
-      return { ...prevState, controlledPosition: { x: paletteX, y: paletteY } };
+    const { t } = useTranslation('sirius-components-diagrams', { keyPrefix: 'palette' });
+
+    const [state, setState] = useState<PaletteState>({
+      searchToolValue: '',
     });
-  }, [paletteX, paletteY]);
 
-  const { setLastToolInvoked, getLastToolInvoked } = useDiagramPalette();
+    const { setLastToolInvoked, getLastToolInvoked } = useDiagramPalette();
 
-  const lastToolInvoked = palette ? getLastToolInvoked(palette.id) : null;
+    const lastToolInvoked = palette ? getLastToolInvoked(palette.id) : null;
 
-  const handleToolClick = (tool: GQLTool) => {
-    onClose();
-    domNode?.focus();
-    onToolClick(tool);
-    if (palette) {
-      setLastToolInvoked(palette.id, tool);
-    }
-    const position = getUpdatedModalPosition({ x: state.controlledPosition.x, y: state.controlledPosition.y }, nodeRef);
-    setState((prevState) => {
-      return { ...prevState, controlledPosition: position };
-    });
-  };
+    const handleToolClick = (tool: GQLTool) => {
+      onClose();
+      onToolClick(tool);
+      if (palette) {
+        setLastToolInvoked(palette.id, tool);
+      }
+    };
 
-  const updatedModalPosition = () => {
-    setState((prevState) => {
-      return {
-        ...prevState,
-        controlledPosition: getUpdatedModalPosition(
-          { x: prevState.controlledPosition.x, y: prevState.controlledPosition.y },
-          nodeRef
-        ),
-      };
-    });
-  };
+    const onSearchFieldValueChanged = (newValue: string): void => {
+      setState((prevState) => ({ ...prevState, searchToolValue: newValue }));
+    };
 
-  useEffect(() => {
-    if (!nodeRef.current) return;
-    // If the palette size changes when opening a section then update the position of the modal
-    const resizeObserver = new ResizeObserver(() => {
-      updatedModalPosition();
-    });
-    resizeObserver.observe(nodeRef.current);
-    updatedModalPosition();
-    return () => resizeObserver.disconnect();
-  }, [paletteX, paletteY]);
+    const handleBackToMainList = () => {
+      if (nodeRef.current) {
+        nodeRef.current.focus();
+      }
+    };
 
-  const draggableBounds = getUpdatedBounds(nodeRef);
-
-  const onPaletteDragStop = (_event, data: DraggableData) => {
-    setState((prevState) => ({ ...prevState, controlledPosition: data }));
-  };
-
-  const onSearchFieldValueChanged = (newValue: string): void => {
-    setState((prevState) => ({ ...prevState, searchToolValue: newValue }));
-  };
-
-  const handleBackToMainList = () => {
-    if (nodeRef.current) {
-      nodeRef.current.focus();
-    }
-  };
-
-  return (
-    <Draggable
-      nodeRef={nodeRef}
-      handle="#tool-palette-header"
-      bounds={draggableBounds}
-      position={state.controlledPosition}
-      onStop={onPaletteDragStop}>
+    return (
       <Paper
-        ref={nodeRef}
+        {...remainingProps}
+        ref={ref}
         data-testid="Palette"
         elevation={3}
         onClick={(event) => event.stopPropagation()}
@@ -209,6 +164,6 @@ export const Palette = ({
           </Box>
         </ClickAwayListener>
       </Paper>
-    </Draggable>
-  );
-};
+    );
+  }
+);
